+++ src/keys.rs
@@ -391,56 +391,38 @@ impl ProofGenerationKey {
     }
 }
 
-#[derive(Debug, Clone)]
-pub struct ViewingKey {
-    pub ak: SpendValidatingKey,
-    pub nk: NullifierDerivingKey,
-}
-
-impl ViewingKey {
-    pub fn rk(&self, ar: jubjub::Fr) -> redjubjub::VerificationKey<SpendAuth> {
-        self.ak.randomize(&ar)
-    }
-
-    pub fn ivk(&self) -> SaplingIvk {
-        SaplingIvk(crh_ivk(self.ak.to_bytes(), self.nk.0.to_bytes()))
-    }
-
-    pub fn to_payment_address(&self, diversifier: Diversifier) -> Option<PaymentAddress> {
-        self.ivk().to_payment_address(diversifier)
-    }
-}
-
-/// A Sapling key that provides the capability to view incoming and outgoing transactions.
-#[derive(Debug)]
+/// A key that provides the capability to view incoming and outgoing transactions.
+///
+/// Modern Zcash wallets use multiple viewing keys scoped to external and internal
+/// operations. You should consider using [`DiversifiableFullViewingKey`] instead, which
+/// handles these details and ensures that you have a full and consistent view of wallet
+/// activity.
+///
+/// Defined in [Zcash Protocol Spec § 3.1: Payment Addresses and Keys][addressesandkeys].
+///
+/// [`DiversifiableFullViewingKey`]: crate::zip32::DiversifiableFullViewingKey
+/// [addressesandkeys]: https://zips.z.cash/protocol/protocol.pdf#addressesandkeys
+// TODO: Rename to `ScopedFullViewingKey` or `ScopedViewingKey` for API clarity.
+#[derive(Clone, Debug, PartialEq, Eq)]
 pub struct FullViewingKey {
-    pub vk: ViewingKey,
-    pub ovk: OutgoingViewingKey,
-}
-
-impl Clone for FullViewingKey {
-    fn clone(&self) -> Self {
-        FullViewingKey {
-            vk: ViewingKey {
-                ak: self.vk.ak.clone(),
-                nk: self.vk.nk,
-            },
-            ovk: self.ovk,
-        }
-    }
+    ak: SpendValidatingKey,
+    nk: NullifierDerivingKey,
+    ovk: OutgoingViewingKey,
 }
 
 impl FullViewingKey {
     pub fn from_expanded_spending_key(expsk: &ExpandedSpendingKey) -> Self {
         FullViewingKey {
-            vk: ViewingKey {
-                ak: (&expsk.ask).into(),
-                nk: (&expsk.nsk).into(),
-            },
+            ak: (&expsk.ask).into(),
+            nk: (&expsk.nsk).into(),
             ovk: expsk.ovk,
         }
     }
 
+    /// Parses a full viewing key from its "raw" encoding as specified in
+    /// [Zcash Protocol Spec § 5.6.3.3: Sapling Full Viewing Keys][saplingfullviewingkeyencoding].
+    ///
+    /// [saplingfullviewingkeyencoding]: https://zips.z.cash/protocol/protocol.pdf#saplingfullviewingkeyencoding
     pub fn read<R: Read>(mut reader: R) -> io::Result<Self> {
         let ak = {
             let mut buf = [0u8; 32];
@@ -471,24 +453,46 @@ impl FullViewingKey {
         reader.read_exact(&mut ovk)?;
 
         Ok(FullViewingKey {
-            vk: ViewingKey { ak, nk },
+            ak,
+            nk,
             ovk: OutgoingViewingKey(ovk),
         })
     }
 
+    /// Serializes the full viewing key as specified in
+    /// [Zcash Protocol Spec § 5.6.3.3: Sapling Full Viewing Keys][saplingfullviewingkeyencoding].
+    ///
+    /// [saplingfullviewingkeyencoding]: https://zips.z.cash/protocol/protocol.pdf#saplingfullviewingkeyencoding
     pub fn write<W: Write>(&self, mut writer: W) -> io::Result<()> {
-        writer.write_all(&self.vk.ak.to_bytes())?;
-        writer.write_all(&self.vk.nk.0.to_bytes())?;
+        writer.write_all(&self.ak.to_bytes())?;
+        writer.write_all(&self.nk.0.to_bytes())?;
         writer.write_all(&self.ovk.0)?;
 
         Ok(())
     }
 
+    /// Serializes the full viewing key as specified in
+    /// [Zcash Protocol Spec § 5.6.3.3: Sapling Full Viewing Keys][saplingfullviewingkeyencoding].
+    ///
+    /// [saplingfullviewingkeyencoding]: https://zips.z.cash/protocol/protocol.pdf#saplingfullviewingkeyencoding
     pub fn to_bytes(&self) -> [u8; 96] {
         let mut result = [0u8; 96];
         self.write(&mut result[..])
             .expect("should be able to serialize a FullViewingKey");
         result
+    }
+
+    pub fn rk(&self, ar: jubjub::Fr) -> redjubjub::VerificationKey<SpendAuth> {
+        self.ak.randomize(&ar)
+    }
+
+    /// Derives an `IncomingViewingKey` for this full viewing key.
+    pub fn ivk(&self) -> SaplingIvk {
+        SaplingIvk(crh_ivk(self.ak.to_bytes(), self.nk.0.to_bytes()))
+    }
+
+    pub fn to_payment_address(&self, diversifier: Diversifier) -> Option<PaymentAddress> {
+        self.ivk().to_payment_address(diversifier)
     }
 }
 
@@ -750,7 +754,7 @@ pub mod testing {
 
     prop_compose! {
         pub fn arb_incoming_viewing_key()(fvk in arb_full_viewing_key()) -> SaplingIvk {
-            fvk.vk.ivk()
+            fvk.ivk()
         }
     }
 }